[Refactor] #18 WaitingPopupView를 WaitingRequestView와 WaitingCompleteView로 분할
