+++ .gitignore
@@ -1,2 +1,9 @@
+## User settings
+xcuserdata/
+*.xcuserstate
+
+## Xcode 8 and earlier
+*.xcscmblueprint
+*.xccheckout
+
 .DS_Store
-*.DS_Store*
