+++ unifest-ios.xcodeproj/xcshareddata/xcschemes/unifest-ios.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1530"
+   LastUpgradeVersion = "1500"
    version = "1.8">
    <BuildAction
       parallelizeBuildables = "YES"
