+++ unifest-ios/Views/Map/SearchSchoolView.swift
@@ -14,63 +14,65 @@ struct SearchSchoolView: View {
     @Environment(\.dismiss) var dismiss
     
     var body: some View {
-        VStack {
-            Spacer()
-                .frame(height: 30)
-            
-            Image(.searchBox)
-                .overlay {
-                    HStack {
-                        TextField("학교/축제 이름을 검색해보세요", text: $searchText)
-                            .font(.system(size: 13))
-                        Image(.searchIcon)
-                    }
-                    .padding(.horizontal, 15)
-                }
-                .padding(.bottom)
-            
-            HStack {
-                Text("검색결과")
-                    .font(.system(size: 12))
-                    .foregroundStyle(.gray)
-                Text("총 2개")
-                    .font(.system(size: 12))
+        GeometryReader { geometry in
+            VStack {
                 Spacer()
-            }
-            .padding(.horizontal)
-            .padding(.bottom, 8)
-            
-            VStack(spacing: 6) {
-                LongSchoolBoxView(schoolName: "건국대학교", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08.")
-                LongSchoolBoxView(schoolName: "건국대학교", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08.")
-            }
-            .padding(.horizontal, 15)
-            .padding(.bottom, 6)
-            
-            Image(.boldLine)
-                .resizable()
-                .scaledToFit()
-                .frame(maxWidth: .infinity)
-            
-            HStack {
-                Text("나의 관심 축제")
-                    .font(.system(size: 15))
-                    .fontWeight(.semibold)
+                    .frame(height: 30)
                 
-                Spacer()
-            }
-            .padding(.horizontal)
-            .padding(.vertical, 10)
-            
-            ScrollView {
-                LazyVGrid(columns: columns) {
-                    /* SchoolBoxView(isSelected: .constant(false), schoolImage: .konkukLogo, schoolName: "건국대 서울캠", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )
-                    SchoolBoxView(isSelected: .constant(false), schoolImage: .hongikLogo, schoolName: "홍익대 서울캠", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )
-                    SchoolBoxView(isSelected: .constant(false), schoolImage: .chungangLogo, schoolName: "중앙대 서울캠", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )
-                    SchoolBoxView(isSelected: .constant(false), schoolImage: .snutLogo, schoolName: "서울과기대", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )
-                    SchoolBoxView(isSelected: .constant(false), schoolImage: .uosLogo, schoolName: "서울시립대", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )*/
+                Image(.searchBox)
+                    .overlay {
+                        HStack {
+                            TextField("학교/축제 이름을 검색해보세요", text: $searchText)
+                                .font(.system(size: 13))
+                            Image(.searchIcon)
+                        }
+                        .padding(.horizontal, 15)
+                    }
+                    .padding(.bottom)
+                
+                HStack {
+                    Text("검색결과")
+                        .font(.system(size: 12))
+                        .foregroundStyle(.gray)
+                    Text("총 2개")
+                        .font(.system(size: 12))
+                    Spacer()
                 }
                 .padding(.horizontal)
+                .padding(.bottom, 8)
+                
+                VStack(spacing: 10) {
+                    LongSchoolBoxView(schoolName: "건국대학교", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08.")
+                    LongSchoolBoxView(schoolName: "건국대학교", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08.")
+                }
+                .padding(.horizontal, 15)
+                .padding(.bottom, 6)
+                
+                Image(.boldLine)
+                    .resizable()
+                    .scaledToFit()
+                    .frame(maxWidth: .infinity)
+                
+                HStack {
+                    Text("나의 관심 축제")
+                        .font(.system(size: 15))
+                        .fontWeight(.semibold)
+                    
+                    Spacer()
+                }
+                .padding(.horizontal)
+                .padding(.vertical, 10)
+                
+                ScrollView {
+                    LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 10), count: 3), spacing: 10) {
+                        SchoolBoxView(isSelected: .constant(false), schoolImageURL: "https://i.namu.wiki/i/frJ2JwLkp_Ts8Le-AVmISJdC2wkhYUte4N5YZ0j-nB_aT_CagQLTOjNDeXMW_kFBeXIe5-pAXDE4CGNVGJ5FMGuS1HUqUTYGfL3AUVJBzum3Ppq5wIUej0v10WuSuOWDT2KIEQwKqA4qiWm3EzS123Uj1gyShvzqaaScOgNsnPg.svg", schoolName: "건국대 서울캠", festivalName: "녹색지대", startDate: "2024-06-24", endDate: "2024-06-27" )
+                        SchoolBoxView(isSelected: .constant(false), schoolImageURL: "https://i.namu.wiki/i/frJ2JwLkp_Ts8Le-AVmISOJycKA-weZiZnZjURRWrVUG1b_6EwEV_XC-T4m0L4tkaLMTrYTt-rg9NsoFAvLcaensvvbxtHJniiJ1ozZQASIXrFBKslQibGRszyoIcpf3Amebt_bZ5rNEYbYSQfkbq21wvQuDNQ5jTqtP1RO3V1g.svg", schoolName: "홍익대 서울캠", festivalName: "녹색지대", startDate: "2024-05-20", endDate: "2024-05-24")
+                        SchoolBoxView(isSelected: .constant(false), schoolImageURL: "https://i.namu.wiki/i/frJ2JwLkp_Ts8Le-AVmISFfwnU-dlgVpQbZ03CrSG0G0U1cfjhp8ygrXEpeJ_s7b0p6hOl8NgkZyZBV-0dSGl1pKcPaPxpTXHqETtFYKDL_QG8Pu1Qa5-79V0ks6ABzw0eOMDa68Hfpunt1HWPZ17GlqMo8zFONfTmVx7HJGfa4.webp", schoolName: "서울대 관악캠", festivalName: "녹색지대", startDate: "2024-06-01", endDate: "2024-06-04")
+                        /* SchoolBoxView(isSelected: .constant(false), schoolImage: .snutLogo, schoolName: "서울과기대", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )
+                         SchoolBoxView(isSelected: .constant(false), schoolImage: .uosLogo, schoolName: "서울시립대", festivalName: "녹색지대", startDate: "05.06.", endDate: "05.08." )*/
+                    }
+                    .padding(.horizontal)
+                }
             }
         }
     }
@@ -94,8 +96,8 @@ struct LongSchoolBoxView: View {
         HStack {
             Image(.konkukLogo)
                 .resizable()
-                .scaledToFit()
                 .frame(width: 52, height: 52)
+                .scaledToFit()
                 .padding(.trailing, 4)
             
             VStack(alignment: .leading) {
@@ -108,6 +110,7 @@ struct LongSchoolBoxView: View {
                 
                 Text(startDate + "-" + endDate)
                     .font(.system(size: 12))
+                    .foregroundStyle(.defaultDarkGray)
             }
             
             Spacer()
